web backend refactoring and randomizer
