+++ web/app.yaml
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 application: muzeiapi
-version: 15
+version: 16
 runtime: python27
 api_version: 1
 threadsafe: false
