+++ web/cron.yaml
@@ -2,3 +2,6 @@ cron:
 - description: update archive daily
   url: /tasks/update_archive
   schedule: every day 02:05
+- description: pick random artwork for next few days if it's empty
+  url: /tasks/pick_random_artwork
+  schedule: every day 02:05
