+++ web/handlers/backroom.py
@@ -20,19 +20,10 @@ import sys
 
 import webapp2
 from google.appengine.ext.webapp import template
-from google.appengine.api import images
-from google.appengine.api import urlfetch
 
-sys.path.append(os.path.join(os.path.dirname(__file__),'../lib'))
-from bs4 import BeautifulSoup
-import cloudstorage as gcs
-
+from handlers import backroomarthelper
 from handlers.common import *
 from models import FeaturedArtwork
-
-
-THUMB_HEIGHT=600
-NO_CROP_TUPLE=(0, 0, 1, 1)
 
 
 def artwork_dict(a):
@@ -82,59 +73,6 @@ class ServiceListHandler(BaseBackroomHandler):
     return json.dumps([artwork_dict(a) for a in queue])
 
 
-def maybe_process_image(image_url, crop_tuple, base_name):
-  if CLOUD_STORAGE_ROOT_URL in image_url and crop_tuple == NO_CROP_TUPLE:
-    return (image_url, None)
-
-  image_result = urlfetch.fetch(image_url, deadline=20)
-  if image_result.status_code < 200 or image_result.status_code >= 300:
-    raise IOError('Error downloading image: HTTP %d.' % image_result.status_code)
-
-  filename = re.sub(r'[^\w]+', '-', base_name.strip().lower()) + '.jpg'
-
-  # main image
-  image_gcs_path = CLOUD_STORAGE_BASE_PATH + '/fullres/' + filename
-  # resize to max width 4000 or max height 2000
-  image_contents = image_result.content
-  image = images.Image(image_contents)
-  edited = False
-  if image.height > 2000:
-    image.resize(width=(image.width * 2000 / image.height), height=2000)
-    edited = True
-  elif image.width > 4000:
-    image.resize(width=4000, height=(image.height * 4000 / image.width))
-    edited = True
-
-  if crop_tuple != NO_CROP_TUPLE:
-    image.crop(*crop_tuple)
-    edited = True
-
-  if edited:
-    image_contents = image.execute_transforms(output_encoding=images.JPEG, quality=80)
-
-  # upload with default ACLs set on the bucket  # or use options={'x-goog-acl': 'public-read'})
-  gcs_file = gcs.open(image_gcs_path, 'w', content_type='image/jpeg')
-  gcs_file.write(image_contents)
-  gcs_file.close()
-
-  # thumb
-  thumb_gcs_path = CLOUD_STORAGE_BASE_PATH + '/thumbs/' + filename
-  thumb = images.Image(image_result.content)
-  thumb.resize(width=(thumb.width * THUMB_HEIGHT / thumb.height), height=THUMB_HEIGHT)
-
-  if crop_tuple != NO_CROP_TUPLE:
-    thumb.crop(*crop_tuple)
-    edited = True
-
-  thumb_contents = thumb.execute_transforms(output_encoding=images.JPEG, quality=40)
-  gcs_file = gcs.open(thumb_gcs_path, 'w', content_type='image/jpeg')
-  gcs_file.write(thumb_contents)
-  gcs_file.close()
-
-  return (CLOUD_STORAGE_ROOT_URL + image_gcs_path,
-          CLOUD_STORAGE_ROOT_URL + thumb_gcs_path)
-
-
 class ServiceAddHandler(BaseBackroomHandler):
   def post(self):
     artwork_json = json.loads(self.request.get('json'))
@@ -147,7 +85,7 @@ class ServiceAddHandler(BaseBackroomHandler):
 
     crop_tuple = tuple(float(x) for x in json.loads(self.request.get('crop')))
 
-    new_image_url, new_thumb_url = maybe_process_image(
+    new_image_url, new_thumb_url = backroomarthelper.maybe_process_image(
         artwork_json['imageUri'],
         crop_tuple,
         publish_date.strftime('%Y%m%d') + ' '
@@ -173,63 +111,10 @@ class ServiceAddFromExternalArtworkUrlHandler(BaseBackroomHandler):
     publish_date = (datetime.datetime
         .utcfromtimestamp(int(self.request.get('publishDate')) / 1000)
         .date())
-    if FeaturedArtwork.all().filter('publish_date =', publish_date).get() != None:
-      webapp2.abort(409, message='Artwork already exists for this date.')
 
-    url = self.request.get('externalArtworkUrl')
-    result = urlfetch.fetch(url)
-    if result.status_code < 200 or result.status_code >= 300:
-      webapp2.abort(400, message='Error processing URL: HTTP %d. Content: %s'
-          % (result.status_code, result.content))
-
-    soup = BeautifulSoup(result.content)
-    attribution = None
-
-    if re.search(r'wikiart.org', url, re.I):
-      attribution = 'wikiart.org'
-      details_url = re.sub(r'#.+', '', url, re.I | re.S) + '?utm_source=Muzei&utm_campaign=Muzei'
-      title = soup.select('h1 span')[0].get_text()
-      author = soup.find(itemprop='author').get_text()
-      completion_year_el = soup.find(itemprop='dateCreated')
-      byline = author + ((', ' + completion_year_el.get_text()) if completion_year_el else '')
-      image_url = soup.find(id='paintingImage')['href']
-    elif re.search(r'metmuseum.org', url, re.I):
-      attribution = 'metmuseum.org'
-      details_url = re.sub(r'[#?].+', '', url, re.I | re.S) + '?utm_source=Muzei&utm_campaign=Muzei'
-      title = soup.find('h2').get_text()
-      author = ''
-      try:
-        author = unicode(soup.find(text='Artist:').parent.next_sibling).strip()
-      except:
-        pass
-      author = re.sub(r'\s*\(.*', '', author)
-      completion_year_el = None
-      try:
-        completion_year_el = unicode(soup.find(text='Date:').parent.next_sibling).strip()
-      except:
-        pass
-      byline = author + ((', ' + completion_year_el) if completion_year_el else '')
-      image_url = soup.find('a', class_='download').attrs['href']
-    else:
-      webapp2.abort(400, message='Unrecognized URL')
-
-    if not title or not author or not image_url:
-      webapp2.abort(500, message='Could not parse HTML')
-
-    image_url, thumb_url = maybe_process_image(image_url,
-        NO_CROP_TUPLE,
-        publish_date.strftime('%Y%m%d') + ' ' + title + ' ' + byline)
-
-    # create the artwork entry
-    new_artwork = FeaturedArtwork(
-        title=title,
-        byline=byline,
-        attribution=attribution,
-        image_url=image_url,
-        thumb_url=thumb_url,
-        details_url=details_url,
-        publish_date=publish_date)
-    new_artwork.save()
+    new_artwork = backroomarthelper.add_art_from_external_details_url(
+        publish_date,
+        self.request.get('externalArtworkUrl'))
 
     self.response.set_status(200)
     self.response.out.write(json.dumps(artwork_dict(new_artwork)))
@@ -248,7 +133,7 @@ class ServiceEditHandler(BaseBackroomHandler):
     target_artwork.byline = artwork_json['byline']
     target_artwork.attribution = artwork_json['attribution'] if 'attribution' in artwork_json else None
 
-    new_image_url, new_thumb_url = maybe_process_image(
+    new_image_url, new_thumb_url = backroomarthelper.maybe_process_image(
         artwork_json['imageUri'],
         crop_tuple,
         target_artwork.publish_date.strftime('%Y%m%d') + ' '
